# ArzTiServer/wwwroot/swagger-original.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Webservice ArzTI API",
    "description": "Kommunikation des Webservers mit den jeweiligen Rechenzentren ",
    "version": "V1"
  },
  "servers": [
    {
      "url": "/v1",
      "description": "Local server"
    },
    {
      "url": "http://arzsoftware.de/v1",
      "description": "Main (production) server"
    },
    {
      "url": "http://staging-api.arzsoftware.de",
      "description": "Internal staging server for testing"
    }
  ],
  "paths": {
...
    "/rezept/abholung": {
      "patch": {
        "tags": [
          "RezeptUIDList"
        ],
        "summary": "Änderung des Abholstatus einer Liste von Rezepten",
        "description": "ändert Abholstatus einer Liste von Rezepten",
        "operationId": "patchRezeptUIdListAbholstatus",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RezeptUId"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RezeptUId"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Returns the newly created item",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Abholstatus"
                  }
                }
              },
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Abholstatus"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": ""
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "BasicAuth": [
              "arz_write"
            ]
          }
        ]
      }
    },
    "/rezept/pruefung": {
      "put": {
        "tags": [
          "RezeptUIDList"
        ],
        "summary": "prüft eine Liste von UUID Rezepten",
        "description": "Vorprüfung einer Liste von UUID Rezepten",
        "operationId": "putRezeptUidListPruefung",
        "requestBody": {
          "description": "optionaler Datensatz",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RezeptUId"
                }
              }
            },
            "application/xml": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RezeptUId"
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "zur Prüfung angenommen"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": ""
          }
        },
        "security": [
          {
            "BasicAuth": [
              "arz_read"
            ]
          }
        ]
      }
    },
    "/rezept/status": {
      "patch": {
        "tags": [
          "RezeptUIDList"
        ],
        "summary": "Statusänderung einer Liste von Rezepten",
        "description": "Statusänderung einer Liste von Rezepten",
        "operationId": "patchRezeptUIdListStatus",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RezeptStatus"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/RezeptStatus"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Liefert den Status einer Liste von Rezepten",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RezeptStatus"
                  }
                }
              },
              "application/*+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RezeptStatus"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": ""
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": ""
          }
        },
        "security": [
          {
            "BasicAuth": [
              "arz_read",
              "arz_write"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "RezeptStatus": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RezeptId"
          },
          "uid": {
            "$ref": "#/components/schemas/RezeptUId"
          },
          "status": {
            "type": "string",
            "description": "Status des Rezepts"
          }
        }
      },
      "RezeptTyp": {
        "type": "string",
        "enum": [
          "ERezept",
          "MRezept",
          "PRezept"
        ]
      },
...
      "RezeptId": {
        "title": "Rezept",
        "type": "object",
        "description": "die Rezept ID",
        "properties": {
          "typ": {
            "$ref": "#/components/schemas/RezeptTyp"
          },
          "id": {
            "type": "string"
          }
        }
      },
      "RezeptUId": {
        "title": "Rezept",
        "type": "object",
        "description": "die Rezept RUID",
        "properties": {
          "ruid": {
            "type": "string"
          }
        }
      },
...
      "Abholstatus": {
        "type": "object",
        "required": [
          "abgeholt"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RezeptId"
          },
          "uid": {
            "$ref": "#/components/schemas/RezeptUId"
          },
          "abgeholt": {
            "type": "boolean"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "NotFound": {
        "description": "The specified resource was not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
...
}